api/requirements.txt
Django
djangorestframework
djangorestframework-jwt
django-rest-authtoken
PyMySQL
mysqlclient
django-cors-headers
drf_yasg
